feat: more reworking of PredictionContext for embedding
