+++ runtime/Go/antlr/v4/transition.go
@@ -131,18 +131,19 @@ var TransitionserializationNames = []string{
 // AtomTransition
 // TODO: make all transitions sets? no, should remove set edges
 type AtomTransition struct {
-	*BaseTransition
+	BaseTransition
 }
 
 func NewAtomTransition(target ATNState, intervalSet int) *AtomTransition {
-	
-	t := new(AtomTransition)
-	t.BaseTransition = NewBaseTransition(target)
-	
-	t.label = intervalSet // The token type or character value or, signifies special intervalSet.
+	t := &AtomTransition{
+		BaseTransition: BaseTransition{
+			target:            target,
+			serializationType: TransitionATOM,
+			label:             intervalSet,
+			isEpsilon:         false,
+		},
+	}
 	t.intervalSet = t.makeLabel()
-	t.serializationType = TransitionATOM
-	
 	return t
 }
 
@@ -161,24 +162,22 @@ func (t *AtomTransition) String() string {
 }
 
 type RuleTransition struct {
-	*BaseTransition
-	
+	BaseTransition
 	followState           ATNState
 	ruleIndex, precedence int
 }
 
 func NewRuleTransition(ruleStart ATNState, ruleIndex, precedence int, followState ATNState) *RuleTransition {
-	
-	t := new(RuleTransition)
-	t.BaseTransition = NewBaseTransition(ruleStart)
-	
-	t.ruleIndex = ruleIndex
-	t.precedence = precedence
-	t.followState = followState
-	t.serializationType = TransitionRULE
-	t.isEpsilon = true
-	
-	return t
+	return &RuleTransition{
+		BaseTransition: BaseTransition{
+			target:            ruleStart,
+			isEpsilon:         true,
+			serializationType: TransitionRULE,
+		},
+		ruleIndex:   ruleIndex,
+		precedence:  precedence,
+		followState: followState,
+	}
 }
 
 func (t *RuleTransition) Matches(_, _, _ int) bool {
@@ -186,20 +185,19 @@ func (t *RuleTransition) Matches(_, _, _ int) bool {
 }
 
 type EpsilonTransition struct {
-	*BaseTransition
-	
+	BaseTransition
 	outermostPrecedenceReturn int
 }
 
 func NewEpsilonTransition(target ATNState, outermostPrecedenceReturn int) *EpsilonTransition {
-	
-	t := new(EpsilonTransition)
-	t.BaseTransition = NewBaseTransition(target)
-	
-	t.serializationType = TransitionEPSILON
-	t.isEpsilon = true
-	t.outermostPrecedenceReturn = outermostPrecedenceReturn
-	return t
+	return &EpsilonTransition{
+		BaseTransition: BaseTransition{
+			target:            target,
+			serializationType: TransitionEPSILON,
+			isEpsilon:         true,
+		},
+		outermostPrecedenceReturn: outermostPrecedenceReturn,
+	}
 }
 
 func (t *EpsilonTransition) Matches(_, _, _ int) bool {
@@ -211,19 +209,20 @@ func (t *EpsilonTransition) String() string {
 }
 
 type RangeTransition struct {
-	*BaseTransition
-	
+	BaseTransition
 	start, stop int
 }
 
 func NewRangeTransition(target ATNState, start, stop int) *RangeTransition {
-	
-	t := new(RangeTransition)
-	t.BaseTransition = NewBaseTransition(target)
-	
-	t.serializationType = TransitionRANGE
-	t.start = start
-	t.stop = stop
+	t := &RangeTransition{
+		BaseTransition: BaseTransition{
+			target:            target,
+			serializationType: TransitionRANGE,
+			isEpsilon:         false,
+		},
+		start: start,
+		stop:  stop,
+	}
 	t.intervalSet = t.makeLabel()
 	return t
 }
@@ -254,37 +253,38 @@ type AbstractPredicateTransition interface {
 }
 
 type BaseAbstractPredicateTransition struct {
-	*BaseTransition
+	BaseTransition
 }
 
 func NewBasePredicateTransition(target ATNState) *BaseAbstractPredicateTransition {
-	
-	t := new(BaseAbstractPredicateTransition)
-	t.BaseTransition = NewBaseTransition(target)
-	
-	return t
+	return &BaseAbstractPredicateTransition{
+		BaseTransition: BaseTransition{
+			target: target,
+		},
+	}
 }
 
 func (a *BaseAbstractPredicateTransition) IAbstractPredicateTransitionFoo() {}
 
 type PredicateTransition struct {
-	*BaseAbstractPredicateTransition
-	
+	BaseAbstractPredicateTransition
 	isCtxDependent       bool
 	ruleIndex, predIndex int
 }
 
 func NewPredicateTransition(target ATNState, ruleIndex, predIndex int, isCtxDependent bool) *PredicateTransition {
-	
-	t := new(PredicateTransition)
-	t.BaseAbstractPredicateTransition = NewBasePredicateTransition(target)
-	
-	t.serializationType = TransitionPREDICATE
-	t.ruleIndex = ruleIndex
-	t.predIndex = predIndex
-	t.isCtxDependent = isCtxDependent // e.g., $i ref in pred
-	t.isEpsilon = true
-	return t
+	return &PredicateTransition{
+		BaseAbstractPredicateTransition: BaseAbstractPredicateTransition{
+			BaseTransition: BaseTransition{
+				target:            target,
+				serializationType: TransitionPREDICATE,
+				isEpsilon:         true,
+			},
+		},
+		isCtxDependent: isCtxDependent,
+		ruleIndex:      ruleIndex,
+		predIndex:      predIndex,
+	}
 }
 
 func (t *PredicateTransition) Matches(_, _, _ int) bool {
@@ -300,23 +300,22 @@ func (t *PredicateTransition) String() string {
 }
 
 type ActionTransition struct {
-	*BaseTransition
-	
+	BaseTransition
 	isCtxDependent                    bool
 	ruleIndex, actionIndex, predIndex int
 }
 
 func NewActionTransition(target ATNState, ruleIndex, actionIndex int, isCtxDependent bool) *ActionTransition {
-	
-	t := new(ActionTransition)
-	t.BaseTransition = NewBaseTransition(target)
-	
-	t.serializationType = TransitionACTION
-	t.ruleIndex = ruleIndex
-	t.actionIndex = actionIndex
-	t.isCtxDependent = isCtxDependent // e.g., $i ref in pred
-	t.isEpsilon = true
-	return t
+	return &ActionTransition{
+		BaseTransition: BaseTransition{
+			target:            target,
+			serializationType: TransitionACTION,
+			isEpsilon:         true,
+		},
+		isCtxDependent: isCtxDependent,
+		ruleIndex:      ruleIndex,
+		actionIndex:    actionIndex,
+	}
 }
 
 func (t *ActionTransition) Matches(_, _, _ int) bool {
@@ -328,22 +327,22 @@ func (t *ActionTransition) String() string {
 }
 
 type SetTransition struct {
-	*BaseTransition
+	BaseTransition
 }
 
 func NewSetTransition(target ATNState, set *IntervalSet) *SetTransition {
-	
-	t := new(SetTransition)
-	t.BaseTransition = NewBaseTransition(target)
-	
-	t.serializationType = TransitionSET
+	t := &SetTransition{
+		BaseTransition: BaseTransition{
+			target:            target,
+			serializationType: TransitionSET,
+		},
+	}
 	if set != nil {
 		t.intervalSet = set
 	} else {
 		t.intervalSet = NewIntervalSet()
 		t.intervalSet.addOne(TokenInvalidType)
 	}
-	
 	return t
 }
 
@@ -356,16 +355,24 @@ func (t *SetTransition) String() string {
 }
 
 type NotSetTransition struct {
-	*SetTransition
+	SetTransition
 }
 
 func NewNotSetTransition(target ATNState, set *IntervalSet) *NotSetTransition {
-	
-	t := new(NotSetTransition)
-	
-	t.SetTransition = NewSetTransition(target, set)
-	
-	t.serializationType = TransitionNOTSET
+	t := &NotSetTransition{
+		SetTransition: SetTransition{
+			BaseTransition: BaseTransition{
+				target:            target,
+				serializationType: TransitionNOTSET,
+			},
+		},
+	}
+	if set != nil {
+		t.intervalSet = set
+	} else {
+		t.intervalSet = NewIntervalSet()
+		t.intervalSet.addOne(TokenInvalidType)
+	}
 	
 	return t
 }
@@ -379,16 +386,16 @@ func (t *NotSetTransition) String() string {
 }
 
 type WildcardTransition struct {
-	*BaseTransition
+	BaseTransition
 }
 
 func NewWildcardTransition(target ATNState) *WildcardTransition {
-	
-	t := new(WildcardTransition)
-	t.BaseTransition = NewBaseTransition(target)
-	
-	t.serializationType = TransitionWILDCARD
-	return t
+	return &WildcardTransition{
+		BaseTransition: BaseTransition{
+			target:            target,
+			serializationType: TransitionWILDCARD,
+		},
+	}
 }
 
 func (t *WildcardTransition) Matches(symbol, minVocabSymbol, maxVocabSymbol int) bool {
@@ -400,21 +407,21 @@ func (t *WildcardTransition) String() string {
 }
 
 type PrecedencePredicateTransition struct {
-	*BaseAbstractPredicateTransition
-	
+	BaseAbstractPredicateTransition
 	precedence int
 }
 
 func NewPrecedencePredicateTransition(target ATNState, precedence int) *PrecedencePredicateTransition {
-	
-	t := new(PrecedencePredicateTransition)
-	t.BaseAbstractPredicateTransition = NewBasePredicateTransition(target)
-	
-	t.serializationType = TransitionPRECEDENCE
-	t.precedence = precedence
-	t.isEpsilon = true
-	
-	return t
+	return &PrecedencePredicateTransition{
+		BaseAbstractPredicateTransition: BaseAbstractPredicateTransition{
+			BaseTransition: BaseTransition{
+				target:            target,
+				serializationType: TransitionPRECEDENCE,
+				isEpsilon:         true,
+			},
+		},
+		precedence: precedence,
+	}
 }
 
 func (t *PrecedencePredicateTransition) Matches(_, _, _ int) bool {
